# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/products/{identifier}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product by ID or slug",
                "description": "Fetch a single product using either its numeric ID or slug.",
                "operationId": "afdab46ef956e2de0edcd9af5455063f",
                "parameters": [
                    {
                        "name": "identifier",
                        "in": "path",
                        "description": "Product ID (numeric) or slug (string)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SingleProductResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - product fetch failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product fetch failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Exception occurred: Something went wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/best-sellers": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Best Seller  products",
                "operationId": "ab4c755e40ecf077b912ebb226b3741e",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit number of products returned (default 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Best Seller  products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductListResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/wise": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get products by type or all types",
                "operationId": "b82da367001c802e19f56a88e7ab6699",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by product type (trending, todays_pick, new_arrival, featured, flash_deal)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "trending"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit number of products returned (default 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Products fetched successfully"
                                        },
                                        "data": {
                                            "description": "Products grouped by type if 'type' not provided",
                                            "properties": {
                                                "trending": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/ProductListResource"
                                                    }
                                                },
                                                "todays_pick": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/ProductListResource"
                                                    }
                                                },
                                                "new_arrival": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/ProductListResource"
                                                    }
                                                },
                                                "featured": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/ProductListResource"
                                                    }
                                                },
                                                "flash_deal": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/ProductListResource"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Exception occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}